add an alert if wishlist's empty

+++ reactapp/src/ScreenMyArticles.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import { Card, Icon, Col, Row } from "antd";
+import { Card, Icon, Col, Row, Alert } from "antd";
 import Nav from "./Nav";
 
 import {connect} from 'react-redux';
@@ -8,25 +8,31 @@ import {connect} from 'react-redux';
 const { Meta } = Card;
 
 function ScreenMyArticles(props) {
-  
-  const articles = props.myArticles.map( (article) => {
-    return  <Col span={8} >
-              <Card style={{marginTop:'2em'}}
-                cover={<img alt={article.wlTitle} src={article.wlImage} />}
-                hoverable={true}
-                actions={[
-                  <Icon type="read" key="ellipsis2" />,
-                  <Icon type="delete" key="ellipsis" 
-                  onClick={ () => props.deleteFromWL(article.wlTitle) } />,
-                ]}
-              >
-                <Meta
-                  title={article.wlTitle}
-                  description={article.wlContent}
-                />
-              </Card>
-            </Col>
-  })
+
+  var articles = []
+
+  if (props.myArticles.length == 0) {
+    articles = <Alert style={{marginTop:'2em'}} message="Pas encore d'article dans vos favoris" type="info" />
+  } else {
+    articles = props.myArticles.map( (article) => {
+      return  <Col span={8} >
+                <Card style={{marginTop:'2em'}}
+                  cover={<img alt={article.wlTitle} src={article.wlImage} />}
+                  hoverable={true}
+                  actions={[
+                    <Icon type="read" key="ellipsis2" />,
+                    <Icon type="delete" key="ellipsis" 
+                    onClick={ () => props.deleteFromWL(article.wlTitle) } />,
+                  ]}
+                >
+                  <Meta
+                    title={article.wlTitle}
+                    description={article.wlContent}
+                  />
+                </Card>
+              </Col>
+    })
+  }
 
   return (
   <div>
